app/requirements.txt
streamlit
pandas
numpy
scikit-learn
catboost
joblib
plotly.express
json
matplotlib.pyplo
datetime
timedelta
BaseEstimator
TransformerMixin
